-Создание компонентов из html разметки (рефакторинг) - Настройка маршрутов
