+++ project/src/components/login-screen/login-screen.tsx
@@ -1,3 +1,6 @@
+import {AppRoute} from '../../const';
+import {Link} from 'react-router-dom';
+
 function LoginScreen (): JSX.Element {
   return (
     <div className="page page--gray page--login">
@@ -5,9 +8,9 @@ function LoginScreen (): JSX.Element {
         <div className="container">
           <div className="header__wrapper">
             <div className="header__left">
-              <a className="header__logo-link" href="main.html">
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
-              </a>
+              <Link className="header__logo-link" to={AppRoute.Main}>
+                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
+              </Link>
             </div>
           </div>
         </div>
@@ -20,7 +23,7 @@ function LoginScreen (): JSX.Element {
             <form className="login__form form" action="#" method="post">
               <div className="login__input-wrapper form__input-wrapper">
                 <label className="visually-hidden">E-mail</label>
-                <input className="login__input form__input" type="email" name="email" placeholder="Email" required/>
+                <input className="login__input form__input" type="email" name="email" placeholder="Email" required />
               </div>
               <div className="login__input-wrapper form__input-wrapper">
                 <label className="visually-hidden">Password</label>
@@ -33,7 +36,7 @@ function LoginScreen (): JSX.Element {
           </section>
           <section className="locations locations--login locations--current">
             <div className="locations__item">
-              <a className="locations__item-link" href="#">
+              <a className="locations__item-link" href="/">
                 <span>Amsterdam</span>
               </a>
             </div>
